+++ common/src/main/java/io/github/mortuusars/envelope/network/packet/clientbound/ClientboundMailboxMenuMailRemovedPacket.java
@@ -3,8 +3,8 @@ package io.github.mortuusars.envelope.network.packet.clientbound;
 import io.github.mortuusars.envelope.Envelope;
 import io.github.mortuusars.mortaar.network.packet.Packet;
 import io.github.mortuusars.envelope.world.inventory.MailboxMenu;
-import io.github.mortuusars.envelope.world.item.component.Id;
 import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.PacketFlow;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
@@ -15,12 +15,12 @@ import org.jetbrains.annotations.NotNull;
 /**
  * Updates the client when mail is removed from the storage by hopper or something.
  */
-public record ClientboundMailboxMenuMailRemovedPacket(Id id) implements Packet {
+public record ClientboundMailboxMenuMailRemovedPacket(int slot) implements Packet {
     public static final ResourceLocation ID = Envelope.resource("mailbox_menu_mail_removed");
     public static final Type<ClientboundMailboxMenuMailRemovedPacket> TYPE = new Type<>(ID);
 
     public static final StreamCodec<RegistryFriendlyByteBuf, ClientboundMailboxMenuMailRemovedPacket> STREAM_CODEC = StreamCodec.composite(
-          Id.STREAM_CODEC, ClientboundMailboxMenuMailRemovedPacket::id,
+          ByteBufCodecs.INT, ClientboundMailboxMenuMailRemovedPacket::slot,
           ClientboundMailboxMenuMailRemovedPacket::new
     );
 
@@ -36,7 +36,7 @@ public record ClientboundMailboxMenuMailRemovedPacket(Id id) implements Packet {
             return false;
         }
 
-        menu.onMailRemoved(id);
+        menu.onMailRemoved(slot);
         return true;
     }
 }